% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickroll_console.R
\name{rickroll_console}
\alias{rickroll_console}
\title{rickroll_console}
\usage{
rickroll_console(delay = 1.5)
}
\arguments{
\item{delay}{numeric delay in seconds between lines.}
}
\value{
Nothing.
}
\description{
Self explanatory. This function rickrolls you in the console.
}
\examples{
rickroll_console(delay=0.5)
}

