% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickroll_youtube.R
\name{rickroll_youtube}
\alias{rickroll_youtube}
\title{rickroll_youtube}
\usage{
rickroll_youtube()
}
\value{
Nothing.
}
\description{
Self explanatory. This function rickrolls you with youtube.
}
\examples{
rickroll_youtube()
}

